Dedupe and clean up
