+++ node/src/chain_specs/manta.rs
@@ -16,6 +16,7 @@
 
 use super::*;
 use crate::command::MANTA_PARACHAIN_ID;
+use manta_primitives::helpers::{get_account_id_from_seed, get_collator_keys_from_seed};
 
 pub type MantaChainSpec = sc_service::GenericChainSpec<manta_runtime::GenesisConfig, Extensions>;
