+++ runtime/primitives/src/lib.rs
@@ -19,6 +19,8 @@
 #![allow(clippy::upper_case_acronyms)]
 #![cfg_attr(not(feature = "std"), no_std)]
 
+pub mod helpers;
+
 pub mod constants;
 pub use constants::time;
